+++ packages/rocket_client/lib/rocket_client.dart
@@ -2,3 +2,4 @@ library rocket_client;
 
 export 'src/rocket_client_base.dart';
 export 'src/extensions.dart';
+export 'src/retry_options.dart';