+++ packages/rocket_mini_view/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
 * [CHANGELOG](https://github.com/JahezAcademy/flutter_rocket/blob/dev/packages/flutter_rocket/CHANGELOG.md)
+
+## 0.0.2
+
+* Fix nullable rocket value issue